Added some dependencies that we will need to implement room database. I also did a small refactor to make the code easier to follow.
