+++ app/src/main/java/com/example/foodapplication/fragments/OutputFragment.kt
@@ -1,14 +1,16 @@
-package com.example.foodapplication
+package com.example.foodapplication.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.*
 import android.widget.TextView
+import com.example.foodapplication.R
 import kotlin.math.roundToInt
 
 class OutputFragment : Fragment() {
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? = inflater.inflate(R.layout.fragment_output, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? = inflater.inflate(
+        R.layout.fragment_output, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
