+++ app/src/main/java/com/example/foodapplication/fragments/ResetPasswordFragment.kt
@@ -1,6 +1,5 @@
-package com.example.foodapplication
+package com.example.foodapplication.fragments
 
-import DatabaseHandler
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
@@ -10,11 +9,15 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import com.example.foodapplication.R
+import com.example.foodapplication.userDatabase.User
+import com.example.foodapplication.userDatabase.DatabaseHandler
 import com.google.android.material.textfield.TextInputLayout
 
 class ResetPasswordFragment : Fragment(), TextWatcher {
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? = inflater.inflate(R.layout.fragment_reset_password, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? = inflater.inflate(
+        R.layout.fragment_reset_password, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
