+++ app/src/main/java/com/example/foodapplication/userDatabase/User.kt
@@ -1,4 +1,4 @@
-package com.example.foodapplication
+package com.example.foodapplication.userDatabase
 
 class User {
