+++ app/src/main/res/layout/fragment_input.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/grey_background"
     android:fillViewport="true"
-    tools:context=".InputFragment">
+    tools:context=".fragments.InputFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
