+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/grey_background"
     android:padding="5dp"
-    tools:context=".LoginFragment">
+    tools:context=".fragments.LoginFragment">
 
     <View
         android:id="@+id/view"
