+++ app/src/main/res/layout/fragment_otp.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/grey_background"
     android:padding="5dp"
-    tools:context=".OtpFragment">
+    tools:context=".fragments.OtpFragment">
 
 
     <EditText
