+++ app/src/main/res/layout/fragment_output.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".OutputFragment">
+    tools:context=".fragments.OutputFragment">
 
     <TextView
         android:id="@+id/totalCalories"
