+++ app/src/main/res/layout/fragment_register.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/grey_background"
     android:padding="5dp"
-    tools:context=".RegisterFragment">
+    tools:context=".fragments.RegisterFragment">
 
     <View
         android:id="@+id/view"
