+++ app/src/main/res/layout/fragment_reset_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/grey_background"
     android:padding="5dp"
-    tools:context=".ResetPasswordFragment">
+    tools:context=".fragments.ResetPasswordFragment">
 
     <TextView
         android:layout_width="wrap_content"
