+++ build.gradle
@@ -9,6 +9,8 @@ buildscript {
         classpath "com.android.tools.build:gradle:4.2.1"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
+        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.4.0-alpha04'
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
